AutoMapperCRUD

+++ MitrosremERP/Controllers/ZaposleniController.cs
@@ -1,11 +1,11 @@
-﻿
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using MitrosremERP.Aplication.Data;
 using MitrosremERP.Aplication.Interfaces;
 using MitrosremERP.Aplication.ViewModels;
 using MitrosremERP.Domain.Models.ZaposleniMitrosrem;
 using MitrosremERP.Infrastructure.Repositories;
+using AutoMapper;
 
 namespace MitrosremERP.Controllers
 {
@@ -13,16 +13,19 @@ namespace MitrosremERP.Controllers
     {
         private readonly IUnitOfWork _unitOfWork;
         private readonly IWebHostEnvironment _webHostEnvironment;
+        private readonly IMapper _autoMapper;
 
-        public ZaposleniController(IUnitOfWork unitOfWork, IWebHostEnvironment webHostEnvironment)
+        public ZaposleniController(IUnitOfWork unitOfWork, IWebHostEnvironment webHostEnvironment, IMapper autoMapper)
         {
             _unitOfWork = unitOfWork;
             _webHostEnvironment = webHostEnvironment;
+            _autoMapper = autoMapper;
         }
         public IActionResult Index()
         {
             IEnumerable<Zaposleni> lista = _unitOfWork.ZaposleniRepository.GetAll();
-            return View(lista);
+            var zaposleniVM = _autoMapper.Map<IEnumerable<ZaposleniVM>>(lista);
+            return View(zaposleniVM);
         }
         [HttpGet]
         public IActionResult Upsert(int? id)
@@ -39,16 +42,36 @@ namespace MitrosremERP.Controllers
                     new SelectListItem { Value = "Musko", Text = "Musko" },
                     new SelectListItem { Value = "Zensko", Text = "Zensko" }
                 },
-                Zaposleni = new Zaposleni()
             };
-           if(id == null || id == 0)
+            if (id == null || id == 0)
             {
                 return View(zaposleniVM);
             }
             else
             {
-                zaposleniVM.Zaposleni = _unitOfWork.ZaposleniRepository.GetFirstOrDefault(i => i.Id == id);
-                return View(zaposleniVM);
+                var zaposleni = _unitOfWork.ZaposleniRepository.GetFirstOrDefault(i => i.Id == id);
+                if (zaposleni == null)
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    var zaposleniVMMapper = _autoMapper.Map<ZaposleniVM>(zaposleni);
+                    {
+                        zaposleniVMMapper.StepenStrucneSpremeLista = _unitOfWork.StepenStrucneSpremeRepository.GetAll().Select(i => new SelectListItem
+                        {
+                            Text = i.StepenObrazovanja,
+                            Value = i.Id.ToString()
+                        });
+                        zaposleniVMMapper.OdaberiPolLista = new List<SelectListItem>
+                        {
+                            new SelectListItem { Value = "Musko", Text = "Musko" },
+                            new SelectListItem { Value = "Zensko", Text = "Zensko" }
+                        };
+                    };
+
+                    return View(zaposleniVMMapper);
+                }
             }
         }
         [HttpPost]
@@ -62,9 +85,9 @@ namespace MitrosremERP.Controllers
                     string filename = Guid.NewGuid().ToString() + Path.GetExtension(file.FileName);
                     var uploads = Path.Combine(imagePath, @"images\zaposleni");
 
-                    if (!string.IsNullOrEmpty(zaposleniVM.Zaposleni.ImageUrl))
+                    if (!string.IsNullOrEmpty(zaposleniVM.ImageUrl))
                     {
-                        var oldImagePath = Path.Combine(imagePath, zaposleniVM.Zaposleni.ImageUrl.TrimStart('\\'));
+                        var oldImagePath = Path.Combine(imagePath, zaposleniVM.ImageUrl.TrimStart('\\'));
                         List<string> images = new List<string>
                         {
                             "user.jpg",
@@ -77,38 +100,49 @@ namespace MitrosremERP.Controllers
                                 System.IO.File.Delete(oldImagePath);
                             }
                         }
-                        
+
                     }
                     using (var fileStreams = new FileStream(Path.Combine(uploads, filename), FileMode.Create))
                     {
                         file.CopyTo(fileStreams);
                     }
-                    zaposleniVM.Zaposleni.ImageUrl = @"\images\zaposleni\" + filename;
+                    zaposleniVM.ImageUrl = @"\images\zaposleni\" + filename;
                 }
                 else
                 {
-                    if(zaposleniVM.Zaposleni.Pol == "Musko")
+                    if (zaposleniVM.Pol == "Musko")
                     {
-                        zaposleniVM.Zaposleni.ImageUrl = @"\images\default\user.jpg";
+                        zaposleniVM.ImageUrl = @"\images\default\user.jpg";
                     }
-                    else if(zaposleniVM.Zaposleni.Pol == "Zensko")
+                    else if (zaposleniVM.Pol == "Zensko")
                     {
-                        zaposleniVM.Zaposleni.ImageUrl = @"\images\default\userW.jpg";
+                        zaposleniVM.ImageUrl = @"\images\default\userW.jpg";
                     }
                 }
-                if (zaposleniVM.Zaposleni.Id == 0)
+                if (zaposleniVM.Id == 0)
                 {
-                    _unitOfWork.ZaposleniRepository.Add(zaposleniVM.Zaposleni);
+                    var zaposleniMapper = _autoMapper.Map<Zaposleni>(zaposleniVM);  
+                    _unitOfWork.ZaposleniRepository.Add(zaposleniMapper);
                     TempData["success"] = "Zaposleni uspesno kreiran";
                 }
                 else
                 {
-                    _unitOfWork.ZaposleniRepository.Update(zaposleniVM.Zaposleni);
+                    var postojiZaposleni = _unitOfWork.ZaposleniRepository.GetFirstOrDefault(i => i.Id == zaposleniVM.Id);
+                    if(postojiZaposleni == null)
+                    {
+                        return NotFound();
+                    }
+                    _autoMapper.Map(zaposleniVM, postojiZaposleni);
                     TempData["success"] = "Uspesno izmenjeni podaci";
+
+                    //_unitOfWork.ZaposleniRepository.Update(zaposleniVM);
                 }
+
                 _unitOfWork.Save();
                 return RedirectToAction("Index");
             }
+
+            ///Popuni podatke dropdown liste ako model nije dobar posle post zahteva
             else
             {
                 zaposleniVM.StepenStrucneSpremeLista = _unitOfWork.StepenStrucneSpremeRepository.GetAll().Select(i => new SelectListItem
@@ -116,6 +150,11 @@ namespace MitrosremERP.Controllers
                     Text = i.StepenObrazovanja,
                     Value = i.Id.ToString()
                 });
+                zaposleniVM.OdaberiPolLista = new List<SelectListItem>
+                        {
+                            new SelectListItem { Value = "Musko", Text = "Musko" },
+                            new SelectListItem { Value = "Zensko", Text = "Zensko" }
+                        };
                 return View(zaposleniVM);
             }
         }
@@ -135,7 +174,7 @@ namespace MitrosremERP.Controllers
                     new SelectListItem { Value = "Musko", Text = "Musko" },
                     new SelectListItem { Value = "Zensko", Text = "Zensko" }
                 },
-                Zaposleni = new Zaposleni()
+               
             };
             if (id == null || id == 0)
             {
@@ -143,23 +182,30 @@ namespace MitrosremERP.Controllers
             }
             else
             {
-                zaposleniVM.Zaposleni = _unitOfWork.ZaposleniRepository.GetFirstOrDefault(i => i.Id == id);
-                return View(zaposleniVM);
+                var zaposleni = _unitOfWork.ZaposleniRepository.GetFirstOrDefault(i => i.Id == id);              
+                var zaposleniVMMapper = _autoMapper.Map<ZaposleniVM>(zaposleni);
+                return View(zaposleniVMMapper);
             }
         }
         [HttpPost]
         public IActionResult DeletePost(int? id)
         {
             var zaposleniObrisi = _unitOfWork.ZaposleniRepository.GetFirstOrDefault(u => u.Id == id);
-            if(zaposleniObrisi == null)
+            if (zaposleniObrisi == null)
             {
                 return NotFound();
             }
             var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, zaposleniObrisi.ImageUrl.TrimStart('\\'));
-            if (System.IO.File.Exists(oldImagePath))
+            List<string> exclusions = new List<string>
+            {  "user.jpg", "userW.jpg" };
+            if (!exclusions.Contains(Path.GetFileName(oldImagePath)))
             {
-                System.IO.File.Delete(oldImagePath);
+                if (System.IO.File.Exists(oldImagePath))
+                {
+                    System.IO.File.Delete(oldImagePath);
+                }
             }
+            
             _unitOfWork.ZaposleniRepository.Remove(zaposleniObrisi);
             _unitOfWork.Save();
             TempData["success"] = "Zaposleni uspesno obrisan";
